requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CMzOugA1hT-4O2kZ5M4Og1VxAfLgir31_besAu0/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
app-store-scraper==0.3.5
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.7
datasets==4.5.0
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fonttools==4.60.2
fpdf2==2.8.4
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~CMzOugA1hT-4O2kZ5M4Og1VxAfLgir31_besAu0/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-generativeai==0.8.6
google-play-scraper==1.2.7
googleapis-common-protos==1.74.0
groq==1.0.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.1
idna==2.10
iniconfig==2.1.0
Jinja2==3.1.6
langdetect==1.0.9
macholib @ file:///AppleInternal/Library/BuildRoots/4~CMzOugA1hT-4O2kZ5M4Og1VxAfLgir31_besAu0/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
numpy==2.0.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.23.0
requests-oauthlib==2.0.0
rich==14.3.3
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/4~CMzOugA1hT-4O2kZ5M4Og1VxAfLgir31_besAu0/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlmodel==0.0.34
starlette==0.49.3
tomli==2.4.0
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.25.11
uvicorn==0.39.0
xxhash==3.6.0
yarl==1.22.0
